+++ examples/03_typescript/src/DisplayRemoteData.tsx
@@ -1,3 +1,5 @@
+/* eslint-disable camelcase */
+
 import * as React from 'react';
 
 import { useFetch } from 'react-hooks-fetch';
@@ -5,10 +7,11 @@ import { useFetch } from 'react-hooks-fetch';
 const Err: React.FC<{ error: Error }> = ({ error }) => <span>Error:{error.message}</span>;
 
 const DisplayRemoteData: React.FC<{ id: string }> = ({ id }) => {
-  const { error, data } = useFetch<{ title: string }>(`https://jsonplaceholder.typicode.com/posts/${id}`);
+  const url = `https://reqres.in/api/users/${id}?delay=3`;
+  const { error, data } = useFetch<{ data: { first_name: string } }>(url);
   if (error) return <Err error={error} />;
   if (!data) return null;
-  return <span>RemoteData:{data.title}</span>;
+  return <div>RemoteData:{data.data.first_name}</div>;
 };
 
 export default DisplayRemoteData;